refactor: move get_selectors to selectors module
