+++ core/src/snapshot/mod.rs
@@ -7,7 +7,6 @@ pub mod util;
 
 use std::{
     collections::{HashMap, HashSet},
-    fs,
     time::Duration,
 };
 
@@ -15,17 +14,15 @@ use clap::{AppSettings, Parser};
 use clap_verbosity_flag::Verbosity;
 use derive_builder::Builder;
 use heimdall_common::{
-    constants::{ADDRESS_REGEX, BYTECODE_REGEX},
     ether::{
         compiler::detect_compiler,
         evm::core::vm::VM,
-        rpc::get_code,
         selectors::{find_function_selectors, resolve_selectors},
-        signatures::{score_signature, ResolvedError, ResolvedFunction, ResolvedLog},
+        signatures::{ResolvedError, ResolvedFunction, ResolvedLog}, bytecode::get_contract_bytecode,
     },
     utils::{
         io::logging::*,
-        strings::{decode_hex, encode_hex_reduced, get_shortned_target},
+        strings::{decode_hex, get_shortned_target},
     },
 };
 use indicatif::ProgressBar;
@@ -34,7 +31,7 @@ use crate::{
     disassemble::{disassemble, DisassemblerArgs},
     snapshot::{
         analyze::snapshot_trace,
-        resolve::match_parameters,
+        resolve::resolve_signatures,
         structures::snapshot::{GasUsed, Snapshot},
         util::tui,
     },
@@ -218,56 +215,7 @@ async fn get_selectors(
     verbose: Verbosity,
     trace: &mut TraceFactory,
     snapshot_call: u32,
-) -> Result<
-    (HashMap<String, u128>, HashMap<String, Vec<ResolvedFunction>>),
-    Box<dyn std::error::Error>,
-> {
-    trace.add_call(
-        snapshot_call,
-        line!(),
-        "heimdall".to_string(),
-        "disassemble".to_string(),
-        vec![format!("{} bytes", contract_bytecode.len() / 2usize)],
-        "()".to_string(),
-    );
-
-    // find and resolve all selectors in the bytecode
-    let disassembled_bytecode = disassemble(DisassemblerArgs {
-        rpc_url,
-        verbose,
-        target: contract_bytecode.to_string(),
-        decimal_counter: false,
-        output: String::new(),
-    })
-    .await?;
-    let selectors = find_function_selectors(evm, &disassembled_bytecode);
-
-    let mut resolved_selectors = HashMap::new();
-    if !skip_resolving {
-        resolved_selectors =
-            resolve_selectors::<ResolvedFunction>(selectors.keys().cloned().collect()).await;
-
-        // if resolved selectors are empty, we can't perform symbolic execution
-        if resolved_selectors.is_empty() {
-            logger.error(&format!(
-                "failed to resolve any function selectors from '{shortened_target}' .",
-                shortened_target = shortened_target
-            ));
-        }
-
-        logger.info(&format!(
-            "resolved {} possible functions from {} detected selectors.",
-            resolved_selectors.len(),
-            selectors.len()
-        ));
-    } else {
-        logger.info(&format!("found {} possible function selectors.", selectors.len()));
-    }
-
-    logger.info(&format!("performing symbolic execution on '{shortened_target}' ."));
-
-    Ok((selectors, resolved_selectors))
-}
+) 
 
 async fn get_snapshots(
     selectors: HashMap<String, u128>,
